Updates to Mock and Component Structure
Lots of reorganization
