+++ .babelrc
@@ -5,6 +5,12 @@
     "stage-0"
   ],
   "plugins": [
-    "add-module-exports"
-  ]
+    "add-module-exports",
+    "transform-decorators-legacy"
+  ],
+  "env": {
+    "development": {
+      "presets": ["react-hmre"]
+    }
+  }
 }
