+++ .eslintrc
@@ -1,42 +1,26 @@
 {
   "parser": "babel-eslint",
-  "plugins": ["react"],
+  "extends": "eslint-config-airbnb",
   "env": {
-    "es6": true,
     "browser": true,
     "node": true,
     "mocha": true
   },
+  "globals": {
+    "connect": true
+  },
   "rules": {
-    "quotes": [2, "single"],
-    "eol-last": [2],
-    "no-mixed-requires": [0],
-    "no-underscore-dangle": [0],
-    "camelcase": 0,
-    "dot-notation": [2, {"allowPattern": "^[a-z]+(_[a-z]+)+$"}],
-    "curly": [2, "multi-line"],
-    "no-unused-expressions": [2, { allowShortCircuit: true, allowTernary: true }],
-    "no-new": 1,
-    "no-unused-vars": 1,
-    "jsx-quotes": [2, "prefer-single"],
-
-    "react/display-name": 2,
-    "react/jsx-boolean-value": 2,
-    "react/jsx-no-undef": 2,
+    "max-len": 0,
     "react/jsx-uses-react": 2,
     "react/jsx-uses-vars": 2,
-    "react/no-did-mount-set-state": 2,
-    "react/no-did-update-set-state": 2,
-    "react/no-multi-comp": 2,
-    "react/no-unknown-property": 2,
-    "react/prop-types": 2,
     "react/react-in-jsx-scope": 2,
-    "react/self-closing-comp": 2,
-    "react/wrap-multilines": 2
+    "block-scoped-var": 0,
+    "padded-blocks": 0,
+    "no-console": 0,
+    "id-length": 0,
+    "no-unused-expressions": 0,
   },
-  "ecmaFeatures": {
-    "jsx": true,
-    "modules": true,
-    "blockBindings": true
-  }
+  "plugins": [
+    "react"
+  ]
 }
