--- app/actions/actionTypes.js
@@ -1,4 +1,0 @@
-export const ADD_HIGHLIGHT = 'ADD_HIGHLIGHT';
-export const ACTIVATE_TOPIC = 'ACTIVATE_TOPIC';
-export const NEW_ARTICLE = 'NEW_ARTICLE';
-export const NEW_QUESTIONS = 'NEW_QUESTIONS';
