+++ app/assets/article_0_mock.json
@@ -1,0 +1,13 @@
+{
+"article_id": 0,
+"text": "Occupy Wall Street supporters gather in Albany                                \nNovember 05, 2011 7:30 am\nBy Steve Lathrop, Albany Democrat-Herald\n\n\n\nBrandishing homemade signs and waving at passing  cars, more than 60 people supporting the Occupy Wall Street movement lined up  in downtown Albany late Friday afternoon to deliver their message.\nAbout five counter-demonstrators were also on hand with signs asking for relief of their tax burden and occasionally debating with individuals from the larger group.\nThe demonstration was peaceful and larger than Edith Orner, one of the organizers, expected.\n“I thought we would get 50 so this is good,” she said.\nOne of those on hand was Nancy Rose-Chism of Albany.\n“A lot of people here are retired, or unemployed or students,” she said, “but I’m not any of those. I wanted to lend my voice to taking back our money and our government.”\nRose-Chism said she was looking for a way to be a part of the growing movement and turned to the Internet.\n“I searched for an Occupy Albany site and found this,” she said. “I’m one of the 99 percent.”\nSo she made her own sign and joined the rally.\nAccording to Orner, many people found out about the demonstration the same way. She said the idea got started among a small group of people who meet once a week. About a week ago they decided to do something in support of the Occupy movements.\n“We just said let’s get on that and we put something together as quickly as we could,” Orner said.\nThat meant placing posters around Albany, and using social networking, emails and the telephone.\nIt seemed to work. By 5 p.m. the crowd had swelled to almost 70.\nThose at the rally lined up on all four corners of Fourth and Ellsworth showing their signs to passing traffic and pedestrians. Orner said the idea was to demonstrate for about an hour.\n“Our concern is what’s going on with the economy,” she said. “We don’t want to reduce the deficit on the backs of those who need (the money) most.”\nRose-Chism said she wanted to be part of something that was trying to make things better for the majority of the people.\n“I believe this is what most of the people want,” she said.\nNo plans for additional demonstrations are in the works now, but that could change.\n“We may have a follow-up discussion of some sort,” Orner said. “We want to be part of doing what we can for the greater good.”\nhttp://democratherald.com/news/local/occupy-wall-street-supporters-gather-in-albany/article_9cae0258-0743-11e1-bc5f-001cc4c002e0.html",
+"date_published": "2011-11-05",
+"city_published": "Albany",
+"state_published": "NY",
+"periodical": "AlbanyDemocratHerald",
+"periodical_code": 1,
+"parse_version": "23",
+"annotators": [
+"SLG"
+]
+}
