+++ app/assets/article_9_mock.json
@@ -1,0 +1,14 @@
+{
+"article_id": 9,
+"text": "3 of 36 DOCUMENTS\nThe Associated Press State & Local Wire\nDecember 22, 2011 Thursday 5:59 PM GMT\nTents still up at Occupy Albany at winter deadline\nSECTION: STATE AND REGIONAL\nLENGTH: 126 words\nDATELINE: ALBANY N.Y.\n \nTents remain pitched at the Occupy Albany encampment on the first day of winter even as the protesters' permit to camp in the city's Academy Park is expiring.\nCity officials have allowed demonstrators to stay overnight in the park near New York's Capitol since Oct. 21 even as Gov. Andrew Cuomo has had state police enforce an 11 p.m. curfew in adjacent Lafayette Park, where more than 130 people have been arrested.\nThe city's chief lawyer told protesters in a letter almost three weeks ago they could stay until Dec 22 under a permit limiting them to 30 tents. It included other conditions with which they have complied. Officials say their concern is the cold. Talks have continued.\nSome protesters say they won't leave, with a rally planned Thursday evening.\n \nLOAD-DATE: December 23, 2011\nLANGUAGE: ENGLISH\nPUBLICATION-TYPE: Newswire\nCopyright 2011 Associated Press\nAll Rights Reserved",
+"date_published": "2011-12-22",
+"city_published": "Albany",
+"state_published": "NY",
+"periodical": "theassociatedpressstatelocalwire",
+"periodical_code": 9,
+"parse_version": "23",
+"annotators": [
+"SLG",
+"MB"
+]
+}
