+++ app/containers/App/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import configureStore from '../store/appStore.js';
+import configureStore from 'store/appStore.js';
 
 const store = configureStore();
+
+/* global styles for app */
+import './styles/app.scss';
 
 export default React.createClass({
   displayName: 'App',
