+++ app/containers/App/styles/_mixins.scss
@@ -9,3 +9,11 @@
   padding-left: initial;
   margin-bottom: initial;
 }
+
+@mixin newspaper {
+  // text-indent: 20px;
+  // Think about em-dashes and more spaces / automatic <p> creation
+  white-space: pre-wrap;
+  line-height: 1.5;
+  font-family: $font-newspaper;
+}
