+++ app/containers/App/styles/_variables.scss
@@ -1,7 +1,12 @@
+// TYPOGRAPHY
 $font-newspaper: georgia, 'PT Serif', serif;
 $font-ui: 'Avenir Next', 'Montserrat', sans-serif;
+
+// ANIMATIONS
 $transition-fast: 0.3s ease;
 $transition-slow: 0.6s ease;
+
+// COLORS
 $offblack      : #333;
 $offwhite      : #ededed;
 $red_accent    : #F16061;
