+++ app/containers/App/styles/app.scss
@@ -1,4 +1,7 @@
-@import 'styleguide';
+// GLOBAL STYLES
+
+@import 'variables';
+@import 'fade';
 
 body {
   font-size: 16px;
