+++ app/styles/CollapsibleList.scss
@@ -1,5 +1,5 @@
-@import 'common';
-@import 'styleguide';
+@import 'app/containers/App/styles/_mixins.scss';
+@import 'app/containers/App/styles/_variables.scss';
 
 .variable-list {
   background: #fff;
