+++ app/styles/TopicPicker.scss
@@ -1,5 +1,5 @@
-@import 'styleguide';
-@import 'common';
+@import 'app/containers/App/styles/_mixins.scss';
+@import 'app/containers/App/styles/_variables.scss';
 
 $topic-picker-nav-width: 25%;
