+++ package.json
@@ -20,10 +20,12 @@
     "babel-eslint": "^4.1.8",
     "babel-loader": "^6.2.1",
     "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.4.3",
     "babel-polyfill": "^6.6.1",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
+    "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.3.13",
     "babel-require": "^1.0.1",
     "babel-runtime": "^6.3.19",
@@ -34,10 +36,11 @@
     "csswring": "^4.2.1",
     "debug": "^2.1.3",
     "del": "^2.2.0",
-    "eslint": "^1.10.3",
-    "eslint-loader": "^1.2.1",
-    "eslint-plugin-react": "^3.16.1",
-    "esprima-fb": "^15001.1001.0-dev-harmony-fb",
+    "eslint": "^2.9.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.9.2",
+    "eslint-plugin-jsx-a11y": "^1.5.3",
+    "eslint-plugin-react": "^5.0.1",
     "exports-loader": "^0.6.2",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.1",
